Correction of an error related to session recovery.
After the session is restored, only requests via the private_request() method will work, since the headers for the request are initialized there via _send_private_request().
in the same requests as, for example, in photo_upload(), requests to self.private() / requests.Session() go directly without initializing headers.

Unfortunately, we had to break the encapsulation rule and initialize the headers in auth.py, otherwise it would be necessary to change all the methods that send requests and work with requests.Session() directly without an intermediary.

Fix #367